java tests almost done

+++ src/utils/config/tester/junit/java.config.js
@@ -6,6 +6,10 @@ const path = require('path');
 const { getInputs, cleanStdout } = require('../../compiler/_utils.js');
 
 const installCommands = {
+  jmock: `curl -0 https://repo1.maven.org/maven2/org/jmock/jmock-junit4/2.12.0/jmock-junit4-2.12.0.jar -o ./.breathecode/jmock.jar`,
+  mockito: `curl -0 https://search.maven.org/remotecontent?filepath=org/mockito/mockito-core/3.2.0/mockito-core-3.2.0.jar -o ./.breathecode/mockito.jar`,
+  objenesis: `curl -0 https://repo1.maven.org/maven2/org/objenesis/objenesis/3.1/objenesis-3.1.jar -o ./.breathecode/objenesis.jar`,
+  bytebuddy: `curl -0 https://search.maven.org/remotecontent?filepath=net/bytebuddy/byte-buddy/1.10.4/byte-buddy-1.10.4.jar -o ./.breathecode/bytebuddy.jar`,
   junit: `curl -0 https://search.maven.org/remotecontent?filepath=junit/junit/4.13-rc-1/junit-4.13-rc-1.jar -o ./.breathecode/junit.jar`,
   hamcrest: `curl -0 https://search.maven.org/remotecontent?filepath=org/hamcrest/hamcrest/2.2/hamcrest-2.2.jar -o ./.breathecode/hamcrest.jar`
 };
@@ -19,6 +23,13 @@ module.exports = (files) => ({
     if (!shell.which('java')) {
       const packageName = "java";
       throw Error(`🚫 You need to have ${packageName} installed to run test the exercises`);
+    }
+
+    if (!fs.existsSync('./.breathecode/mockito.jar')){
+      if(!this.install('mockito') || !this.install('bytebuddy') || !this.install('objenesis')){
+        Console.error("There was a problem instaling mockito");
+        throw new Error("There was a problem instaling mockito");
+      }
     }
 
     if (!fs.existsSync('./.breathecode/junit.jar')){
@@ -51,22 +62,22 @@ module.exports = (files) => ({
     const appPath = files.map(f => './'+f.path).find(f => f.indexOf('App.java') > -1);
 
     const content = fs.readFileSync(appPath, "utf8");
-    const count = getInputs(/input\((?:["'`]{1}(.*)["'`]{1})?\)/gm, content);
+    const count = getInputs(/reader\.readLine\((?:["'`]{1}(.*)["'`]{1})?\)/gm, content);
     let answers = (count.length == 0) ? [] : await socket.ask(count);
 
     const rootPath = this.getEntryPath().replace('Test.java', '');
     const cmd = `
-      javac -cp ./.breathecode/junit.jar ${this.getEntryPath()} ${appPath} &&
-      java -cp ${rootPath}:./.breathecode/junit.jar:./.breathecode/hamcrest.jar org.junit.runner.JUnitCore Test
+      javac -cp ./.breathecode/mockito.jar:./.breathecode/junit.jar ${this.getEntryPath()} ${appPath} &&
+      java -cp ${rootPath}:./.breathecode/hamcrest.jar:./.breathecode/objenesis.jar:./.breathecode/bytebuddy.jar:./.breathecode/mockito.jar:./.breathecode/junit.jar org.junit.runner.JUnitCore Test
     `;
     return cmd
   },
   cleanup: async function(socket){
     const rootPath = this.getEntryPath().replace('Test.java', '');
-    const cmd = `
+    if (fs.existsSync(`${rootPath}*.class`)) return `
       rm ${rootPath}*.class
     `;
-    return cmd
+    else return null;
   }
 
 });
